medconnect-backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
django-filter==23.5
django-storages==1.14.2
Pillow==10.2.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0 